pyproject.toml
[project]
name = "evaluation-apiserver"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "qingru97", email = "qingru97@gmail.com"},
]
dependencies = [
    "uvicorn>=0.30.1",
    "fastapi>=0.111.1",
    "python-multipart>=0.0.9",
    "elasticsearch==8.14.0",
    "rapidfuzz==3.9.4",
    "deepeval==0.21.67",
    "llama-cpp-python==0.2.82",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
start = "python ./api/api.py --reload"
working_dir = "."